Naive Sample

+++ quintet/Defs.h
@@ -4,12 +4,14 @@
 #include <cstdint>
 
 #include <string>
+#include <functional>
 
 // typedefs
 namespace quintet {
 
 using PrmIdx = std::uint64_t;
-using Port   = std::uint8_t;
+using Port   = std::uint16_t;
+
 
 }
